pyproject.toml
[tool.poetry]
name = "selenium-env"
version = "0.1.0"
description = ""
authors = ["johnkdo <johnkdo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "<3.0"
notebook = "^6.0.3"
django-extensions = "^2.2.9"
selenium = "^3.141.0"
beautifulsoup4 = "^4.8.2"
requests = "^2.23.0"
pillow = "^7.0.0"
django-phone-field = "^1.8.0"
django-phonenumber-field = "^4.0.0"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
